requirements.txt
huggingface_hub
keras
numpy
Pillow
streamlit
tensorflow

